% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cavaan-package.R
\docType{package}
\name{cavaan-package}
\alias{cavaan}
\alias{cavaan-package}
\title{cavaan: Latent Variable Analysis using Structural Equation Models (SEM)}
\description{
Estimation of structural equation models (SEM).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/Kss2k/cavaan}
}

}
\author{
\strong{Maintainer}: Kjell Solem Slupphaug \email{slupphaugkjell@gmail.com} (\href{https://orcid.org/0009-0005-8324-2834}{ORCID})

}
\keyword{internal}
